refactor: common AwareScrollView implementation
