+++ example/src/components/AwareScrollView/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./KeyboardAwareScrollView";
