+++ example/src/components/AwareScrollView/useSmoothKeyboardHandler.ts
