+++ example/src/screens/Examples/AwareScrollView/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useResizeMode } from 'react-native-keyboard-controller';
 
-import KeyboardAwareScrollView from './KeyboardAwareScrollView';
+import KeyboardAwareScrollView from '../../../components/AwareScrollView';
 import TextInput from './TextInput';
 import { styles } from './styles';
 
@@ -9,7 +9,7 @@ export default function AwareScrollView() {
   useResizeMode();
 
   return (
-    <KeyboardAwareScrollView style={styles.container} contentContainerStyle={styles.content}>
+    <KeyboardAwareScrollView bottomOffset={50} style={styles.container} contentContainerStyle={styles.content}>
       {new Array(10).fill(0).map((_, i) => (
         <TextInput
           key={i}
